New grant expiry interval static
Because previously the phpleague/oauth library used `DateTime::class` object any time a modification was done to the timestamp, the `DateTime::class` would happily return the modified object value. But then PHPLeague corrected their library to start using the `DateTimeImmutable::class.

Thus when our getters and setters were used for the intervals on our token expiry they would always try and calculate based on the time `now()`.  
Which is problematic because that means all token would expire immediate... i.e. `now() - now() = 0`.

Unfortunately as there is no configuration or magical methods on the grant itself to set the access token expiry and to avoid at all costs to forking the dependency library. This required some way of globally setting the access token expiry. Thus opted to set a configurable static that could be reused via the controller (really a service at this point) to allow for this.

Making it static so if people wanted to override within their project.

Happy days. For now.