Refactor of POC features
Not required to assign a member for Confidential grants. Ideally a new feature request for assigning members should be applied to this module.

Because this module also has its own DB fields for confidential the EntityTrait class was not able to set and get our Confidential fields thus getter and setters written in this change.